move ModeModal into components
